Fixed critical timing bug

+++ src/core/Scene.h
@@ -58,7 +58,7 @@ class Scene {
     const std::string name;
     SceneData data;
     double timeDelta = 0.0;
-    long lastTimeNs = 0L;
+    long long lastTimeNs = 0L;
 
     unsigned int width = 0;
     unsigned int height = 0;
